Move the ethereum address field to the main module and clean up the code.
