--- ethereum_address_field/composer.json
@@ -1,14 +1,0 @@
-{
-  "name": "drupal/ethereum_address_field",
-  "type": "drupal-module",
-  "description": "Ethereum 20 digit hex address field.",
-  "keywords": ["Drupal"],
-  "license": "GPL-2.0+",
-  "homepage": "https://www.drupal.org/project/ethereum_address_field",
-  "minimum-stability": "dev",
-  "support": {
-    "issues": "https://www.drupal.org/project/issues/ethereum_address_field",
-    "source": "http://cgit.drupalcode.org/ethereum_address_field"
-  },
-  "require": { }
-}
