--- ethereum_address_field/ethereum_address_field.module
@@ -1,32 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains ethereum_address_field.module.
- */
-
-use Drupal\Core\Routing\RouteMatchInterface;
-
-/**
- * Implements hook_help().
- */
-function ethereum_address_field_help($route_name, RouteMatchInterface $route_match) {
-  switch ($route_name) {
-    // Main module help for the ethereum_address_field module.
-    case 'help.page.ethereum_address_field':
-      return ethereum_get_readme_html('ethereum_address_field');
-    default:
-  }
-}
-
-/**
- * Implements hook_theme().
- */
-function ethereum_address_field_theme() {
-  return [
-    'ethereum_address_field' => [
-      'template' => 'ethereum_address_field',
-      'render element' => 'children',
-    ],
-  ];
-}
