--- ethereum_address_field/src/Plugin/Field/FieldFormatter/EthereumAddressFormatterType.php
@@ -1,80 +1,0 @@
-<?php
-
-namespace Drupal\ethereum_address_field\Plugin\Field\FieldFormatter;
-
-use Drupal\Component\Utility\Html;
-use Drupal\Core\Field\FieldItemInterface;
-use Drupal\Core\Field\FieldItemListInterface;
-use Drupal\Core\Field\FormatterBase;
-use Drupal\Core\Form\FormStateInterface;
-
-/**
- * Plugin implementation of the 'ethereum_address_formatter_type' formatter.
- *
- * @FieldFormatter(
- *   id = "ethereum_address_formatter_type",
- *   label = @Translation("Ethereum address formatter type"),
- *   field_types = {
- *     "ethereum_address"
- *   }
- * )
- */
-class EthereumAddressFormatterType extends FormatterBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function defaultSettings() {
-    return [
-      // Implement default settings.
-    ] + parent::defaultSettings();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsForm(array $form, FormStateInterface $form_state) {
-    return [
-      // Implement settings form.
-    ] + parent::settingsForm($form, $form_state);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function settingsSummary() {
-    $summary = [];
-    // Implement settings summary.
-
-    return $summary;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function viewElements(FieldItemListInterface $items, $langcode) {
-    $elements = [];
-
-    foreach ($items as $delta => $item) {
-      $elements[$delta] = ['#markup' => $this->viewValue($item)];
-    }
-
-    return $elements;
-  }
-
-  /**
-   * Generate the output appropriate for one field item.
-   *
-   * @param \Drupal\Core\Field\FieldItemInterface $item
-   *   One field item.
-   *
-   * @return string
-   *   The textual output generated.
-   */
-  protected function viewValue(FieldItemInterface $item) {
-    // The text value has no text format assigned to it, so the user input
-    // should equal the output, including newlines.
-    return nl2br(Html::escape($item->value));
-  }
-
-}
