--- ethereum_address_field/templates/ethereum_address_field.html.twig
@@ -1,1 +1,0 @@
-<!-- Add you custom twig html here -->
