+++ ethereum_user_connector/config/optional/core.entity_form_display.user.user.default.yml
@@ -8,7 +8,6 @@ dependencies:
     - field.field.user.user.user_picture
     - image.style.thumbnail
   module:
-    - ethereum_address_field
     - ethereum_user_connector
     - image
     - user
@@ -31,7 +30,7 @@ content:
     settings: {  }
     third_party_settings: {  }
   field_ethereum_address:
-    type: ethereum_address_widget
+    type: ethereum_address
     weight: 2
     settings:
       size: 22
